added iddevice script to obtain id
