+++ register.py
@@ -5,31 +5,10 @@ import sqlite3
 import json
 import uuid
 import time
+import iddevice
 
-def getID():
-    conn = sqlite3.connect('rbpi-rmit-iot.db')
-    c = conn.cursor()
-    c.execute("CREATE TABLE IF NOT EXISTS info (id TEXT PRIMARY KEY)")
-
-    c.execute("SELECT id FROM info")
-    fetch = c.fetchone();
-    FETCH_ID = fetch[0] if fetch != None else None;
-    
-    if(FETCH_ID == None):
-        ID=str(uuid.uuid4()).replace('-','')
-        print('initialise' + ID)
-        c.execute("INSERT INTO info (id) VALUES (?)", [ID])
-    else:
-        ID = FETCH_ID
-
-    conn.commit()
-    conn.close()
-
-    return ID
-
-def register(id):
-    ID = id
-    print("Hello")
+def register():
+    ID = iddevice.getID()
     # URL of the endpoint
     url = "https://yeotf6yzv3.execute-api.ap-southeast-2.amazonaws.com/default/RegisterDevice"
     
@@ -49,5 +28,4 @@ def register(id):
             continue
         break
 
-ID = getID()
-register(ID)
+register()
